chore: migrate splashtinder
This reverts commit 7acdf501ce6adae380f090d198aef052b4e96868.
