+++ src/components/setting/View.tsx
@@ -12,12 +12,11 @@ import SettingListItem from './helpers/SettingListItem';
 import LanguageSettings from './LanguageSettings';
 import AboutSettings from './AboutSettings';
 import SplashSettings from './SplashSettings';
-import SplashSettingsR from './SplashSettingsReanimated';
 import DownloadSettings from './DownloadSettings';
 import LoginSettings from '../login/View';
 import PremiumSettings from '../billing/View';
 import SponsorBlockSettings from './sponsorblock/View';
-import { isAndroid, isIOS, isOldArch } from '@utils/RNUtils';
+import { isAndroid, isIOS } from '@utils/RNUtils';
 import FlexView from '@components/commonui/FlexViewNewArch';
 import DefaultScreenOption from '@enums/ScreenOption';
 
@@ -130,11 +129,6 @@ const HomeSettings = ({ navigation }: Props) => {
   );
 };
 
-const SplashView = () => {
-  if (isOldArch() && isAndroid) return <SplashSettings />;
-  return <SplashSettingsR />;
-};
-
 const Settings = ({ headerBackVisible = true }: Props) => {
   const { t } = useTranslation();
 
@@ -151,7 +145,7 @@ const Settings = ({ headerBackVisible = true }: Props) => {
         />
         <Stack.Screen
           name={NoxView.SPLASH_GALLARY}
-          component={SplashView}
+          component={SplashSettings}
           options={{
             title: t('Settings.SplashSettingName'),
             ...DefaultScreenOption,
